requirements.txt
fastapi
uvicorn
mysql-connector-python
pymongo
transformers
torch
python-dotenv
